+++ USC Films/ViewModel/TV Shows/TVShowViewModel.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct TVShowViewModel {
-    var id : Float
+    var id : Int
     var title: String
     var rating: String
     var releaseYear: String
